tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-play-state_running.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-play-state: running</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-play-state-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="When a paused animation is resumed, it restarts
   from the current value, not necessarily from the beginning of the animation">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes move {
         0%, 24% {
            left: 0px;
            top: 0px;
         }
         25%, 49% {
            left: 100px;
            top: 0px;
         }
         50%, 74% {
            left: 100px;
            top: 100px;
         }
         75%, 100% {
            left: 0px;
            top: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
         animation-duration: 1s;
      }
      .test {
         animation-name: move;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var style = getComputedStyle(div, null);
         var passed = false;
         setTimeout(function() {
            div.style.setProperty('animation-play-state', 'paused', '');
            passed = (style.left == '100px' && style.top == '0px');
         }, 370);
         setTimeout(function() {
            div.style.setProperty('animation-play-state', 'running', '');
         }, 870);
         setTimeout(function() {
            passed = (passed && style.left == '100px' && style.top == '100px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         }, 1120);
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue block below should jump rightwards, pause for a while, then jump
   downwards, then leftwards, then upwards to the initial position.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
